requirements.txt
aiogram==3.4.1
aiogram-dialog==2.1.0
python-dotenv==1.0.1
